add shardingshpere usage example to cosid-example-shardingshpere
