+++ examples/cosid-example-shardingsphere/build.gradle.kts
@@ -49,10 +49,17 @@ dependencies {
     implementation(project(":cosid-spring-boot-starter"))
     implementation(project(":cosid-jackson"))
 
+    //region cosid-mybatis
+    implementation(project(":cosid-mybatis"))
+    implementation("org.mybatis.spring.boot:mybatis-spring-boot-starter")
+    //endregion
+
     //region cosid-jdbc
     implementation(project(":cosid-jdbc"))
     implementation("org.springframework.boot:spring-boot-starter-jdbc")
     implementation("mysql:mysql-connector-java")
+    implementation(project(":cosid-shardingsphere"))
+    implementation("org.apache.shardingsphere:shardingsphere-jdbc-core-spring-boot-starter:${rootProject.ext.get("shardingsphereVersion")}")
     //endregion
 
     implementation("org.springframework.boot:spring-boot-starter-actuator")
